+++ Sources/AutomergeSwiftAdditions/Codable/Encoding/AutomergeSingleValueEncodingContainer.swift
@@ -2,6 +2,7 @@ import struct Automerge.Counter
 import class Automerge.Document
 import struct Automerge.ObjId
 import protocol Automerge.ScalarValueRepresentable
+import enum Automerge.Value
 import Foundation
 
 struct AutomergeSingleValueEncodingContainer: SingleValueEncodingContainer {
@@ -181,26 +182,44 @@ struct AutomergeSingleValueEncodingContainer: SingleValueEncodingContainer {
             // Capture and override the default encodable pathing for Counter since
             // Automerge supports it as a primitive value type.
             let downcastText = value as! Text
-            // FIXME: check to see if the object exists here before just splatting a new one into place
 
-            let textNode: ObjId
+            let existingValue: Value?
+            // get any existing value - type of `get` based on the key type
             if let indexToWrite = codingkey.intValue {
-                textNode = try document.putObject(obj: objectId, index: UInt64(indexToWrite), ty: .Text)
+                existingValue = try document.get(obj: objectId, index: UInt64(indexToWrite))
             } else {
-                textNode = try document.putObject(obj: objectId, key: codingkey.stringValue, ty: .Text)
+                existingValue = try document.get(obj: objectId, key: codingkey.stringValue)
+            }
+
+            let textNodeId: ObjId
+            if let existingNode = existingValue {
+                guard case let .Object(textId, .Text) = existingNode else {
+                    throw CodingKeyLookupError
+                        .mismatchedSchema(
+                            "Text Encoding on KeyedContainer at \(self.codingPath) exists and is \(existingNode), not Text."
+                        )
+                }
+                textNodeId = textId
+            } else {
+                // no existing value is there, so create a Text node
+                if let indexToWrite = codingkey.intValue {
+                    textNodeId = try document.putObject(obj: objectId, index: UInt64(indexToWrite), ty: .Text)
+                } else {
+                    textNodeId = try document.putObject(obj: objectId, key: codingkey.stringValue, ty: .Text)
+                }
             }
 
             // Iterate through
-            let currentText = try! document.text(obj: textNode).utf8
+            let currentText = try! document.text(obj: textNodeId).utf8
             let diff: CollectionDifference<String.UTF8View.Element> = downcastText.value.utf8
                 .difference(from: currentText)
             for change in diff {
                 switch change {
                 case let .insert(offset, element, _):
                     let char = String(bytes: [element], encoding: .utf8)
-                    try document.spliceText(obj: textNode, start: UInt64(offset), delete: 0, value: char)
+                    try document.spliceText(obj: textNodeId, start: UInt64(offset), delete: 0, value: char)
                 case let .remove(offset, _, _):
-                    try document.spliceText(obj: textNode, start: UInt64(offset), delete: 1)
+                    try document.spliceText(obj: textNodeId, start: UInt64(offset), delete: 1)
                 }
             }
         default:
